fix(inventory): stock.committed na transacao da baixa, disponibilidade e snapshot do autor
O stock.committed era gravado depois da transacao de commit, e a outbox so aceita
escrita dentro de transacao (Propagation.MANDATORY): a primeira entrega baixava o
estoque e estourava, e a reentrega caia em ALREADY_COMMITTED sem publicar. O pedido
ficava pago para sempre. A escrita passou para dentro de StockCommitTransaction.

Disponibilidade retornava max(0, -reservado), ignorando o estoque do produto.

O cliente do user recebia o ServiceTokenProvider mas nao mandava o Authorization:
a avaliacao era gravada sem nome nem foto do autor (401 silencioso).

Tambem: rate da avaliacao como short para casar com SMALLINT, coluna inexistente
em ReviewEligibilityEntity e value-serializer do produtor para o DLT.

+++ micro-services/inventory/src/main/java/ecommerce_event_driven/inventory/modules/review/infra/outbound/repos/ReviewJpaRepository.java
@@ -45,7 +45,7 @@ public interface ReviewJpaRepository extends JpaRepository<ReviewEntity, Long> {
      * Busca reviews nao deletadas por rate.
      */
     Page<ReviewEntity> findByIdProductAndRateAndDeletedAtIsNullOrderByCreatedAtDesc(
-            Long idProduct, Integer rate, Pageable pageable);
+            Long idProduct, Short rate, Pageable pageable);
 
     /**
      * Busca todas as reviews do usuario nao deletadas.
@@ -85,7 +85,7 @@ public interface ReviewJpaRepository extends JpaRepository<ReviewEntity, Long> {
                and r.rate = :rate
                and r.deletedAt is null
             """)
-    Long countByProductAndRate(@Param("idProduct") Long idProduct, @Param("rate") Integer rate);
+    Long countByProductAndRate(@Param("idProduct") Long idProduct, @Param("rate") Short rate);
 
     /**
      * Soft delete de uma review.
@@ -106,7 +106,7 @@ public interface ReviewJpaRepository extends JpaRepository<ReviewEntity, Long> {
             """)
     int updateReview(
             @Param("id") Long id,
-            @Param("rate") Integer rate,
+            @Param("rate") Short rate,
             @Param("title") String title,
             @Param("description") String description);
