+++ micro-services/inventory/src/main/java/ecommerce_event_driven/inventory/modules/review/infra/outbound/repos/entity/ReviewEntity.java
@@ -39,7 +39,7 @@ public class ReviewEntity {
     private Long idOrder;
 
     @Column(name = "rate", nullable = false)
-    private Integer rate;
+    private Short rate;
 
     @Column(name = "title", nullable = false, length = 150)
     private String title;
